requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
SQLAlchemy==2.0.43
PyMySQL==1.1.1
python-dotenv==1.1.1
redis==5.2.1
kafka-python==2.2.15
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
pytest==8.3.4
